# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sable Operator API
  version: 1.2.0
  description: |
    Loopback-only operator API for the Sable C2 server. Bound to
    `127.0.0.1:8443` by default. All endpoints except `POST /api/auth/login`
    require a Bearer JWT obtained from `/api/auth/login`.

    **Error responses are plain text** (`text/plain; charset=utf-8`), not JSON.
    Each error response below documents the canonical message body emitted by
    the server.

    The TLS certificate is self-signed and pinned by SHA-256 fingerprint when
    operators connect through `sablectl`. Browsers will need to accept the
    certificate manually.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: https://127.0.0.1:8443
    description: Operator API on the Sable server (loopback)
security:
  - bearerAuth: []
...
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Exchange operator password for a JWT
      description: |
        Returns a signed HS256 JWT valid for 1 hour. Rate-limited to 5
        attempts per 60 seconds per remote IP; rate-limited responses
        include a `Retry-After` header.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              password: correct-horse-battery-staple
      responses:
        '200':
          description: Authentication succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '413': { $ref: '#/components/responses/RequestEntityTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429':
          description: Too many login attempts from this IP.
          headers:
            Retry-After:
              schema: { type: integer }
              description: Seconds until another attempt is allowed.
          content:
            text/plain:
              schema: { type: string }
              example: "too many requests\n"
  /api/audit:
    get:
      tags: [Audit]
      operationId: listAuditEvents
      summary: Recent operator and session audit events
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Up to 512 most recent events.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditEvent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
  /api/admin/shutdown:
    post:
      tags: [Admin]
      operationId: shutdownServer
      summary: Gracefully stop the local Sable server
      description: |
        Requests the same graceful shutdown path used for an operating-system
        termination signal. The operator API is loopback-only and this route
        requires a valid operator JWT.
      responses:
        '202':
          description: Shutdown accepted.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '503':
          description: The host process did not configure lifecycle control.
  /api/overview:
    get:
      tags: [Fleet]
      operationId: getFleetOverview
      summary: Lightweight fleet status and activity summary
      description: Returns dashboard counters, one summary row per registered agent, bounded task-outcome buckets, and recent semantic activity without output or artifact payloads.
      responses:
        '200':
          description: Current fleet overview.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FleetOverview' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
...
  /api/agents:
    get:
      tags: [Agents]
      operationId: listAgents
      summary: List registered agents
      description: |
        Returns the full agent list. Each entry omits `outputs` (use
        `GET /api/agents/{id}` to fetch task output history) and never
        includes the agent's pre-shared secret.
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Agent list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Agent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }

    post:
      tags: [Agents]
      operationId: registerAgent
      summary: Pre-register an agent session
      description: |
        Creates the server-side agent record so an implant built with the
        same `id` and `secret_hex` can begin beaconing. Idempotent: re-registering
        an existing ID preserves notes, tags, queued tasks, outputs, and artifacts.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterAgentRequest' }
      responses:
        '201':
          description: Agent registered.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                required: [id]
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '413': { $ref: '#/components/responses/RequestEntityTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
...
  /api/agents/{id}/rekey:
    parameters:
      - $ref: '#/components/parameters/AgentIdParam'
    post:
      tags: [Agents]
      operationId: rekeyAgent
      summary: Generate and install a new 32-byte agent secret
      description: The new secret is returned once and the currently deployed agent must be rebuilt or reconfigured.
      responses:
        '200':
          description: Secret rotated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RekeyAgentResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
...
  /api/agents/{id}/artifacts:
    parameters:
      - $ref: '#/components/parameters/AgentIdParam'
    get:
      tags: [Artifacts]
      operationId: listArtifacts
      summary: List server-side artifact summaries
      description: Response items omit `data`; fetch a single artifact to retrieve its base64 contents.
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Artifact summaries (no `data`).
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Artifact' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
    post:
      tags: [Artifacts]
      operationId: saveArtifact
      summary: Save an operator artifact
      description: |
        Maximum body size is ~75 MB. If `key` is supplied and matches an
        existing artifact for this agent, the existing record is returned
        (idempotent save).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Artifact' }
      responses:
        '201':
          description: Artifact saved (response omits `data`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Artifact' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '413': { $ref: '#/components/responses/RequestEntityTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
  /api/agents/{id}/artifacts/{artifactID}:
    parameters:
      - $ref: '#/components/parameters/AgentIdParam'
      - $ref: '#/components/parameters/ArtifactIdParam'
    get:
      tags: [Artifacts]
      operationId: getArtifact
      summary: Fetch a single artifact including base64 data
      responses:
        '200':
          description: Full artifact including `data`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Artifact' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
    delete:
      tags: [Artifacts]
      operationId: deleteArtifact
      summary: Delete a retained artifact and its data blob
      responses:
        '204': { description: Artifact deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
  /api/agents/{id}/artifacts/retention:
    parameters:
      - $ref: '#/components/parameters/AgentIdParam'
    put:
      tags: [Artifacts]
      operationId: setArtifactRetention
      summary: Retain only the newest configured number of artifacts
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ArtifactRetentionRequest' }
      responses:
        '204': { description: Retention updated and oldest excess artifacts removed. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '413': { $ref: '#/components/responses/RequestEntityTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 JWT issued by `/api/auth/login`. Issuer `sable-operator`,
        audience `sable-web`, subject `operator`, 1-hour TTL.
  parameters:
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 500, default: 500 }
      description: Maximum number of collection items to return.
    OffsetParam:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
      description: Zero-based collection offset. Responses include X-Total-Count, X-Limit, and X-Offset headers.
    AgentIdParam:
      name: id
      in: path
      required: true
      description: Agent identifier.
      schema:
        type: string
        pattern: '^[a-zA-Z0-9-]{1,64}$'
        minLength: 1
        maxLength: 64
        example: main
    ArtifactIdParam:
      name: artifactID
      in: path
      required: true
      description: Artifact identifier.
      schema:
        type: string
        pattern: '^[a-zA-Z0-9-]{1,64}$'
        minLength: 1
        maxLength: 64
...
  schemas:
    LoginRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          description: Operator password. Request body capped at 4096 bytes.
          example: correct-horse-battery-staple
    LoginResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: Bearer JWT (HS256). Valid for 1 hour.
    RegisterAgentRequest:
      type: object
      required: [id, secret_hex]
      properties:
        id:
          type: string
          pattern: '^[a-zA-Z0-9-]{1,64}$'
          minLength: 1
          maxLength: 64
          example: main
        secret_hex:
          type: string
          description: 32-byte agent secret encoded as 64 lowercase hex characters.
          pattern: '^[0-9a-fA-F]{64}$'
          minLength: 64
          maxLength: 64
        display_name:
          type: string
          maxLength: 64
          description: Initial operator-facing name, normally seeded from the local build label.
    RekeyAgentResponse:
      type: object
      required: [id, secret_hex]
      properties:
        id: { type: string }
        secret_hex:
          type: string
          pattern: '^[0-9a-f]{64}$'
    ArtifactRetentionRequest:
      type: object
      required: [max_items]
      properties:
        max_items: { type: integer, minimum: 1, maximum: 256 }
...
    FleetOverview:
      type: object
      required: [generated_at, total, never_seen, on_schedule, overdue, offline, retired, queued_tasks, running_tasks, active_jobs, active_transfers, failed_last_24_hours, agents, task_outcomes_24h, task_outcomes_7d, recent_activity, failure_alerts]
      properties:
        generated_at: { type: string, format: date-time }
        total: { type: integer, minimum: 0 }
        never_seen: { type: integer, minimum: 0 }
        on_schedule: { type: integer, minimum: 0 }
        overdue: { type: integer, minimum: 0 }
        offline: { type: integer, minimum: 0 }
        retired: { type: integer, minimum: 0 }
        queued_tasks:
          type: integer
          minimum: 0
          description: Tasks waiting for first delivery to a host.
        running_tasks:
          type: integer
          minimum: 0
          description: Tasks delivered to a host and awaiting a terminal result.
        active_jobs:
          type: array
          items: { $ref: '#/components/schemas/ActiveJob' }
        active_transfers: { type: integer, minimum: 0 }
        failed_last_24_hours: { type: integer, minimum: 0 }
        agents:
          type: array
          items: { $ref: '#/components/schemas/AgentSummary' }
        task_outcomes_24h:
          type: array
          description: Twenty-four hourly buckets, oldest first.
          minItems: 24
          maxItems: 24
          items: { $ref: '#/components/schemas/TaskOutcomeBucket' }
        task_outcomes_7d:
          type: array
          description: Seven daily buckets, oldest first.
          minItems: 7
          maxItems: 7
          items: { $ref: '#/components/schemas/TaskOutcomeBucket' }
        recent_activity:
          type: array
          maxItems: 32
          items: { $ref: '#/components/schemas/FleetActivityEvent' }
        failure_alerts:
          type: array
          maxItems: 64
          description: Unresolved failed-task notifications from the last 24 hours.
          items: { $ref: '#/components/schemas/FailureAlert' }
...
    AuditEvent:
      type: object
      required: [timestamp, action]
      properties:
        timestamp:
          type: string
          format: date-time
        agent_id:
          type: string
          description: Empty for events that aren't tied to a specific agent.
        action:
          type: string
          example: queue_task
        detail:
          type: string
  responses:
    BadRequest:
      description: Malformed request, validation failure, or invalid task type.
      content:
        text/plain:
          schema: { type: string }
          example: "bad request\n"
    Unauthorized:
      description: Missing or invalid Bearer token.
      content:
        text/plain:
          schema: { type: string }
          example: "unauthorized\n"
    NotFound:
      description: Agent, task, or artifact does not exist.
      content:
        text/plain:
          schema: { type: string }
          example: "404 page not found\n"
    MethodNotAllowed:
      description: Method not supported on this path.
      content:
        text/plain:
          schema: { type: string }
          example: "method not allowed\n"
    RequestEntityTooLarge:
      description: Request body exceeded the per-endpoint byte cap.
      content:
        text/plain:
          schema: { type: string }
          example: "request body too large\n"
    UnsupportedMediaType:
      description: Body Content-Type is not `application/json`.
      content:
        text/plain:
          schema: { type: string }
          example: "content type must be application/json\n"